requirements.txt
Flask==1.1.2
ibm-watson
python-telegram-bot
